--- ioccheck/__init__.py
@@ -1,3 +1,0 @@
-from .hash import Hash
-from .ioc import IOC, IOCReport
-from .ip import IP
