+++ ioccheck/iocs/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+""" isort:skip_file """
+from .ioc import IOC, IOCReport
+from .hash import Hash, HashReport
+from .ip import IP, IPReport
