+++ ioccheck/services/__init__.py
@@ -1,3 +1,7 @@
+#!/usr/bin/env python
+""" isort:skip_file """
+
+from .service import Service
 from .malwarebazaar import MalwareBazaar
 from .shodan import Shodan
 from .virustotal import VirusTotal
