+++ ioccheck/services/malwarebazaar.py
@@ -6,7 +6,7 @@ import requests
 from backoff import expo, on_exception
 from ratelimit import RateLimitException, limits
 
-from ioccheck.services.service import Service
+from ioccheck.services import Service
 
 logger = logging.getLogger(__name__)
